+++ src/routes/users.js
@@ -1,6 +1,7 @@
 import { Router } from 'express'
 import { Users } from '../db/models/index.js'
 import { userData } from '../data/userData.js'
+import { invalidIdError, badRequestError, notFoundError } from '../data/errorMessages.js'
 
 const usersRouter = Router()
 
@@ -34,7 +35,9 @@ usersRouter.post('/bulkcreate', async (req, res, next) => {
 usersRouter.route('/:userId')
 .get(async (req, res, next) => {
     try {
+        if (req.params.userId.length !== 36) return next({ code: 400, msg: invalidIdError })
         const user = await Users.findByPk(req.params.userId)
+        if (!user) return next({ code: 404, msg: notFoundError })
         res.send(user)
     } catch (error) {
         next(error)
@@ -42,10 +45,12 @@ usersRouter.route('/:userId')
 })
 .put(async (req, res, next) => {
     try {
+        if (req.params.userId.length !== 36) return next({ code: 400, msg: invalidIdError })
         const user = await Users.update(req.body, {
             where: { id: req.params.userId },
             returning: true
         })
+        if (!user) return next({ code: 404, msg: notFoundError })
         res.send(user[1][0])
     } catch (error) {
         next(error)
@@ -53,9 +58,11 @@ usersRouter.route('/:userId')
 })
 .delete(async (req, res, next) => {
     try {
+        if (req.params.userId.length !== 36) return next({ code: 400, msg: invalidIdError })
         const result = await Users.destroy({
             where: { id: req.params.userId }
         })
+        if (!result) return next({ code: 404, msg: notFoundError })
         res.sendStatus(204)
     } catch (error) {
         next(error)
